# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Onboarding API
  version: 0.10.0
  description: 'API MVP-платформы интерактивного онбординга: управление проектами, элементами интерфейса, сценариями
    и шагами, Runtime API для SDK, Tracking, аналитика и PDF-отчёты.'
servers:
  - url: http://localhost:8080
    description: Локальное окружение
...
paths:
...
  /api/v1/projects/{projectId}/scenarios:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags:
        - Scenarios
      operationId: createScenario
      summary: Создать сценарий
      security:
        - AdminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenarioRequest'
      responses:
        '201':
          description: Сценарий создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
      description: Создаёт сценарий в статусе in_development.
    get:
      tags:
        - Scenarios
      operationId: listScenarios
      summary: Получить список сценариев проекта
      security:
        - AdminBearer: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ScenarioStatus'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список сценариев
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Возвращает сценарии проекта с фильтрацией по статусу и пагинацией.
  /api/v1/scenarios/{scenarioId}:
    parameters:
      - $ref: '#/components/parameters/ScenarioId'
    get:
      tags:
        - Scenarios
      operationId: getScenario
      summary: Получить сценарий
      security:
        - AdminBearer: []
      responses:
        '200':
          description: Сценарий найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioWithSteps'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Возвращает сценарий вместе с упорядоченным списком шагов.
    patch:
      tags:
        - Scenarios
      operationId: updateScenario
      summary: Изменить сценарий
      description: Изменение разрешено для сценариев в статусах in_development и disabled. Сценарий в статусе enabled
        изменять нельзя.
      security:
        - AdminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScenarioRequest'
      responses:
        '200':
          description: Сценарий обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags:
        - Scenarios
      operationId: deleteScenario
      summary: Удалить сценарий
      security:
        - AdminBearer: []
      responses:
        '204':
          description: Сценарий мягко удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      description: Мягко удаляет сценарий в любом статусе. Удалённый сценарий перестаёт возвращаться в списках и Runtime,
        но исторические сессии, события и аналитика сохраняются.
  /api/v1/scenarios/{scenarioId}/publish:
    parameters:
      - $ref: '#/components/parameters/ScenarioId'
    post:
      tags:
        - Scenarios
      operationId: publishScenario
      summary: Опубликовать сценарий
      description: Переводит сценарий из in_development в enabled.
      security:
        - AdminBearer: []
      responses:
        '200':
          description: Сценарий опубликован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/scenarios/{scenarioId}/enable:
    parameters:
      - $ref: '#/components/parameters/ScenarioId'
    post:
      tags:
        - Scenarios
      operationId: enableScenario
      summary: Включить сценарий
      security:
        - AdminBearer: []
      responses:
        '200':
          description: Сценарий включён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
      description: Переводит сценарий из disabled в enabled.
  /api/v1/scenarios/{scenarioId}/disable:
    parameters:
      - $ref: '#/components/parameters/ScenarioId'
    post:
      tags:
        - Scenarios
      operationId: disableScenario
      summary: Выключить сценарий
      security:
        - AdminBearer: []
      responses:
        '200':
          description: Сценарий выключен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      description: Переводит сценарий из enabled в disabled. После выключения сценарий не выдаётся Runtime и может быть
        отредактирован.
...
components:
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      description: Административный токен для MVP.
    ProjectKey:
      type: apiKey
      in: header
      name: X-Project-Key
      description: Публичный ключ проекта для SDK.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ScenarioId:
      name: scenarioId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Максимальное количество записей в ответе.
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Количество записей, которые нужно пропустить.
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Отсутствует или недействителен административный токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Операция конфликтует с текущим состоянием ресурса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Нарушены правила валидации или бизнес-правила
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: scenario_invalid_state
        message:
          type: string
          example: published scenario cannot be edited
        details:
          type: object
          additionalProperties: true
...
    ScenarioStatus:
      type: string
      enum:
        - in_development
        - enabled
        - disabled
      description: in_development — сценарий настраивается и ещё не публиковался; enabled — доступен Runtime и недоступен
        для изменения; disabled — недоступен Runtime и может быть отредактирован.
    CreateScenarioRequest:
      type: object
      required:
        - name
        - page_pattern
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          default: ''
        page_pattern:
          type: string
          minLength: 1
          maxLength: 2048
          example: /advertisements/create
    UpdateScenarioRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        page_pattern:
          type: string
          minLength: 1
          maxLength: 2048
    Scenario:
      type: object
      required:
        - id
        - project_id
        - name
        - description
        - page_pattern
        - status
        - published_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        page_pattern:
          type: string
        status:
          $ref: '#/components/schemas/ScenarioStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
          nullable: true
          description: Время первой публикации сценария. Равно null, если сценарий ещё ни разу не публиковался.
    ScenarioSummary:
      type: object
      required:
        - id
        - project_id
        - name
        - page_pattern
        - status
        - steps_count
        - published_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        page_pattern:
          type: string
        status:
          $ref: '#/components/schemas/ScenarioStatus'
        steps_count:
          type: integer
          minimum: 0
        published_at:
          type: string
          format: date-time
          nullable: true
          description: Время первой публикации сценария. Равно null, если сценарий ещё ни разу не публиковался.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ScenarioListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioSummary'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
...
    Step:
      type: object
      required:
        - id
        - scenario_id
        - element_id
        - step_num
        - title
        - description
        - frontend_data
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        scenario_id:
          type: string
          format: uuid
        step_num:
          type: integer
          minimum: 1
        title:
          type: string
        description:
          type: string
        frontend_data:
          $ref: '#/components/schemas/FrontendStepData'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        element_id:
          type: string
          format: uuid
          description: Элемент интерфейса проекта, на который ориентирована подсказка.
      description: Шаг онбординга. Содержит ссылку на Element, к которому SDK привязывает подсказку.
...
    FrontendStepData:
      type: object
      description: Произвольная конфигурация шага для frontend/SDK. Точная структура будет согласована с frontend позднее.
        В PostgreSQL поле хранится как jsonb.
      additionalProperties: true
      default: {}
...
    ScenarioWithSteps:
      allOf:
        - $ref: '#/components/schemas/Scenario'
        - type: object
          required:
            - steps
          properties:
            steps:
              type: array
              description: Упорядоченный список шагов сценария.
              items:
                $ref: '#/components/schemas/Step'
      description: Сценарий вместе с полным упорядоченным списком шагов.
...